[FIX] custompopup 파일로정리(#47)
